Lithium integration, pathing fix, item names.
Integrates with Lithium, which now upgrades inventories to LithiumInventories.
Storage blocks now non-pathable, which prevents villagers and chickens getting stuck on the edges of hoppers with linked storages on top.
Item names of handhelds now show in inventory guis. (would be better to tie names to channels, but that is complex...)